test: refactor snapshot #5
