+++ deno.json
@@ -39,7 +39,9 @@
     "exclude": ["**/*_test.*", "src/__OLD/**", "*.todo", "**/tests/**"]
   },
   "imports": {
-    "@fresh/test-utils": "./packages/test-utils/src/mod.ts",
+  "@fresh/test-utils": "./packages/test-utils/src/mod.ts",
+  "@fresh/internal/test-utils": "./packages/internal/src/test-utils.ts",
+  "@fresh/internal/versions": "./packages/internal/src/versions.ts",
     "@deno/doc": "jsr:@deno/doc@^0.172.0",
     "@deno/esbuild-plugin": "jsr:@deno/esbuild-plugin@^1.2.0",
     "@fresh/build-id": "jsr:@fresh/build-id@^1.0.0",
