+++ packages/internal/src/browser.ts
@@ -1,0 +1,1 @@
+export * from "../../test-utils/src/browser.ts";
