+++ packages/internal/src/dom.ts
@@ -1,0 +1,1 @@
+export * from "../../test-utils/src/dom.ts";
