+++ packages/internal/src/fs.ts
@@ -1,0 +1,1 @@
+export * from "../../test-utils/src/fs.ts";
