+++ packages/plugin-vite/tests/build_test.ts
@@ -1,5 +1,5 @@
 import { expect } from "@std/expect";
-import { waitFor, waitForText, withBrowser } from "@fresh/test-utils";
+import { waitFor, waitForText, withBrowser } from "@fresh/internal/test-utils";
 import {
   buildVite,
   DEMO_DIR,
