+++ packages/plugin-vite/tests/dev_server_test.ts
@@ -1,6 +1,6 @@
 import * as path from "@std/path";
 import { expect } from "@std/expect";
-import { waitFor, waitForText, withBrowser } from "@fresh/test-utils";
+import { waitFor, waitForText, withBrowser } from "@fresh/internal/test-utils";
 import {
   DEMO_DIR,
   FIXTURE_DIR,
