+++ www/main_test.ts
@@ -1,4 +1,4 @@
-import { withBrowser, withChildProcessServer } from "@fresh/test-utils";
+import { withBrowser, withChildProcessServer } from "@fresh/internal/test-utils";
 import { expect } from "@std/expect";
 import { retry } from "@std/async/retry";
 import {
